esformat: avoid unnecessary O=>T=>O translation in video_format_TransformTo
video_format_TransformTo called video_format_TransformBy, resulting in an
unnecessary O=>T=>O translation between orientation and transform variants.
I have fixed this by simply moving the format modification logic to *To,
with *By now calling *To instead of the other way around.

Note: this has no impact on functionality; as can be seen with the test
suite; the results of calling these two functions individually is the
same.