esformat: split orient+transform calc to new function
The logic for applying a transformation to an orientation to calculate the
resulting orientation has been moved out of video_format_TransformBy into
a new function - vlc_video_orient_Transform - allowing this calculation to
be done without involving a format object, if desired.

The video_format_TransformBy function itself is now a one liner and so has
been moved to a static inline function within the header.