esformat: fix transformation logic
As demonstrated by the orientation/transform test suite addition, the
logic of some of the relevant functions is flawed, resulting in incorrect
orientations/transformations being calculated. The video_format_TransformTo
function is an exception, where its use of two of the other functions in
combination happens to always give the correct result.

This commit reworks the logic, simplifying it in doing so, correcting the
problems and now allowing the test suite to pass fully.