core: set thread priority options to volatile
It is pointless in having them saved to the saved setting file (and
available in the preferences GUI) since the vlc_threads_setup call
during init occurs too early to take a saved setting into account.

Note, an alternate patch was offered up which changed the point during
init at which the vlc_threads_setup call is made, which caused some
early processing (e.g. saved settings file reading) to occur with
normal priority, but allowed the saved settings to be obeyed, but
that was rejected as apparently the current point at which it occurs
is the desired behaviour; hence this alternate patch to avoid
pointlessly saving these options and presenting them in the GUI.