core: add range for --verbose
Enforce lower bound of zero, and enforce sensible upper bound. The upper
bound of the standard vlc_log_type enum constants is 3, however when this
change was originally discussed the issue of some plugins wanting values
beyond this range for extended debug levels was raised, so a few extra
levels are allowed for here.