deinterlace: convert MMXEXT only accelerations to SSE2
(prep work for purging MMX/MMXEXT)

the code changed here has no greater acceleration than MMX/MMXEXT, so to
prepare for the MMX/MMXEXT purge, convert this to SSE2

note, this is a basic conversion only, *using* SSE2 vector registers (Xmm)
but not actually taking advantage of their greater width; that may
require significant revision to the code and thus is left for a separate
commit (and/or person) to possibly tackle later. note the following
benefits, vs. leaving as it was:
 - switching to SSE registers avoids the MMX<->FP register clash
 - thus it avoids the need for issuing `emms` instructions
 - with a little more work, could take full advantage of these wider registers
 - allows us to proceed with the purge of all old MMX/MMXEXT code

also, I inlined the asm, removing dependency on the abstraction macro set