deinterlace: finish up SSE2 enhancement
In a previous patch, bits of deinterlace that were MMX/MMXEXT accelerated
only were upgraded to SSE2 in preparation for a complete purge of
MMX/MMXEXT, however this upgrade was minimal, not going as far as actually
making proper use of the wider SSE2 registers; todo notes were left
with regard to this.

This commit addresses those todo notes.

With algo-x, the todos were simply removed; in that case using more of the
SSE2 registers would involve moving from the 8x8 processing model to a
16x16/8x16 one, which would then change the results of processing, per how
it determines which change to apply to each 8x8 block on the content of
each 8x8 block.