anaglyph: avoid use with incompatible widths
nothing was stopping this filter being casually being applied to videos
with the correct format, but a width not a multiple of 4; doing so could
result in output that contained green diagonal bands due to the algorithm
not being suited to avoiding this.

the easiest solution is just to block use with incompatible widths.