cmdline: deprecate --nofoo style options
boolean (flag) options until now have supported TWO styles of negative
variants, --no-foo (as advertised in the help output) and also --nofoo.

both forms have been around in VLC for a long time (2002), with initial bool
option support using --foo and --no-foo added in
0170377dc07808be9105b199c5e26142e52a2cd9 and the additional --nofoo form
added a few days later in e058a31531c7d1375a0bc1a3540d4f15cb40da4e
(seemingly for the 0.4.1 release).

only the --no-foo negative form (afaik) has ever been advertised in user
help output though, so unlikely users are even aware of the --nofoo form.

VLC has a LOT of options; prior to my commits removing pre-v3.0 obsolete
ones, a few commits back, with three long options per bool (--foo, --no-foo
and --nofoo) there were in the region of 2400 total unique long options
on my system. The removal of those obsolete options (note that those
that became obsolete in 3.x remain), cut this down by about 200+. If we
remove support for --nofoo, this cuts this down further by about 300. A
saving of about 20%.

This vast number of options to handle in command line argument parsing
does not bring any noticeable delay in loading VLC on my modern system,
however, that does not mean that the wasteful inefficiency from such an
unnecessary (imo) feature is desirable.

Consider that one of the most common usage models for VLC is double-
clicking on a video in a file manager, possibly watching some portion
of it, before closing VLC, followed possibly by many repeats of this
with other videos; this inefficiency is wasting some amount of energy,
or battery life (if that's all you care about) on every instance. Note
that if the user refrains from closing the initial instance when switching
to a different video file, it makes no difference since of course a new
instance of VLC is spun up in the background, detecting and passing the
playback request to the other instance.

support for this secondary form of negative booleans is completely
unnecessary, i am not aware of it being in use, and imo is just pointlessly
burning energy across VLC's millions of user's machines for no benefit.

i propose the following three step process for removal of --nofoo support,
for which this patch completes the first:

 1) [4.0] issue warning of pending future deprecation in response to use
    of --nofoo style negative booleans, while allowing them to remain
    functional (done here).
 2) [5.0?] disable functionality, while issuing a deprecation warning,
    just as done typically for obsoleted options.
 3) [6.0?] fully purge all artifacts.