plugins: add missing cat+subcat before options
a handful of modules were missing set cat+subcat calls before their options,
which causes a problem with respect to displaying them in the GUI preferences
(but not help output).

this commit adds the missing set calls to all of them, but using only the now
standardised "hidden" cat/subcat for the time being; assigning the correct
cat/subcat in each case will be addressed by module-specific follow up commits.

note that the way the Qt GUI works (not certain on the others) options are
not organised on a per category basis; the first category and subcategory
specified are used to select the position in the tree to create an entry
for the plugin, from which all of the plugin's options are available. If
no cat/subcat is given then all options are ignored.

on that basis:
 - the use of the "hidden" cat/subcat makes no difference to visibility of
   options.
 - in the case of decklink, which breaks up its option set under multiple
   subcats (though all lumped under the first, as described above), i have
   left the new set_subcategory() entry, added at the start, disabled, in
   order to avoid changing in this commit where that plugin's entry in the
   tree is located.

this commit serves to get the calls themselves in place, and to highlight
the set of plugins suffering from this issue.