opengl/display: fix option categorisation for gles2
note: the lack of cat/subcat on the gles2 module may have been
deliberate (the hidden cat/subcat use was added by myself a few
commits back), due to the fact that gl and gles2 are separate
plugins from the same source and with this correction they
introduce duplicate options when both are installed on a system,
which is not ideal, however there is a similar issue with vout
plugins and the fact that the options are garuanteed identical
means that it should not be a real issue. This also ensures that
the gles2 plugin will work correctly independant of the gl one
(if such a situation applies to anyone).