disable unnecessary plugin option set cat calls
(most) plugins without options do not need to make calls to set
cat/subcat, but many do, which is pointless inefficiency, so let's
address that.

I am aware that the core has a variant of the module selection config
types which are category based rather than capability based; thus where
that is used, it is important that modules retain such entries. there
were three such modules where this applied, and I have left the calls
enabled, leaving a note against them instead.

i've moved the lines themselves down to the bottom in many cases as part
of a general reorganisation of plugin descriptors, to be expanded upon
in a follow up commit.