mp4: replace use of cat hint with section
mp4 is the only module using cat hint except for the core, and with the
core it is used in effect to specify headings for help text output (the
category stuff is ignored there, unlike the (Qt at least) GUI.

being a cat hint means that this "hacks" sub-heading is printed like a
module heading, which is imperfect from a neatness point of view, but also,
and more importantly, potentially confusing. this is not so if we use
a section, which is likely more what was really intended anyway.