plugins: add 'info' option type
to be used for info-request type options such as --help, which currently
use the boolean type which forces them to have negative variants which are
not really wanted.

they are always volatile, which also implies always private

fyi, an alternate implementation using a boolean subtype (i.e. a type
0x61) was considered, which would not have required the previous commit
expanding the type field from 8bit to 16bit, but it required various
hacks to change the normal boolean type handling in various places.