plugins: refactor capabilities 1/4
This is the first step in revising module capabilities to be largely
enum based instead of string based, in the interest of efficiency (and
because it's cleaner and less prone to mistakes). Note that whilst most
modules will now use an enum based capability, string based will still
be available for rare and/or custom capabilities.

here we:
 - establish the enum defining the set of "built-in" capabilities
 - expand modules_t to hold a new 'capability' enum property, using
   up some padding space
 - update the plugin cache save mechanism to capture the new property
   (and only save/load the old string one when a module has a custom
   capability)
 - deprecate the old module_get_capability() function, which now redirects
   to a helper
 - add two new functions in place of module_get_capability()
    - vlc_module_get_capability(): the primary one, returning an enum
    - vlc_module_get_custom_capability(): for fetching the custom cap
      string for those where the former function returns VLC_CAP_CUSTOM
 - prepare a few tiny pieces of the core for the new design
 - establish a few helper functions for translating between the new
   enum constants and the old string form (a description form is also
   thrown in for potential display purposes)
 - add the vlc_module_get_capability_str() and
   vlc_module_get_capability_name() convenience helpers

capabilities left as custom:
 - "opengl es2"
 - "opencv internal filter"

for now:
 - we have yet to convert modules to using the new enum IDs
 - we have yet to provide a means for them to do so
 - all requests for modules are still done via strings
 - the module bank's capability tree is still organised using the
   string form, so no optimisation gain there yet

@reviewers: I am not firmly committed to the selection of which pile
('built-in' vs. 'custom') each capability belongs in; I just went with my
gut on what seemed a somewhat sensible starting point. By all means shuffle
some around if you really wish, the changes to do so will be trivial.