help: fix module help for mods with no options
it confused me in the past when I tried `vlc -m pulse` and it came back
with an error stating that no matching module was found, and to look in
the list generated by --list, when that does in fact contain a module with
that name.

the problem is that the module has no options (nor pulsesrc/pulselist),
and the underlying logic is geared primarily towards --full-help/--longhelp
output which deliberately skips over such modules.

i have finally looked into this and fixed the confusing behaviour by
avoiding skipping over modules when in search mode, so now we output
minimal details about such modules along with a statement that it has no
options when looking up modules, but continue to skip over them in
--full-help/--longhelp output.