help: avoid printing range in simple [0;n] cases
printing an numberic type's range was already skipped when the range was
the default type min/max, however there are lots of such options that
only accept absolute values, setting zero as the lower bound and leaving
the upper bound as the type max, and for which help output is messier than
desirable both with an arguably unnecessary range output and also the large
value of INT_MAX/FLT_MAX.

this change additionally avoids printing the range in such cases, giving
cleaner and just as useful output

yes, we do nothing to explicitly show whether or not an option is absolute,
i.e. is zero or INT_MIN/-FLT_MAX, but it should be typically obvious from
context, and current logic is that out of range values just get silently
capped anyway.