tests: add unique option checker
this analyses the complete list of all available long and short options,
hunting for conflicting descriptions (duplicates) in order to highlight
potential problems.

currently it is blocked from actually accessing private (GUI-hidden) and
obsolete ones since they get filtered out by module_config_get(); this will
be addressed shortly.

it outputs details of all options and potential problems to stderr
(available in the test's log file), before potentially failing.

originally the test was intended to fail simply should any duplicates be
found at all, however it turns out that a handful of plugins are
deliberately duplicating some or all of their options, so instead, where
duplicates are found, certain important details (like type and int/float
range) are compared and the test only failed if such details differ; thus
where duplication is necessary, it is only a problem (causing test failure)
if the important bits differ. this means that, for instance, if one
plugin declares a fixed range on an integer option, and expects this to be
enforced, while another declares the option without, which could potentially
result in the first handling values outside of the expected range, this
issue will be caught by this test, whereas if they both enforced the same
range, that would be fine and the test would pass.

note that for booleans, both positive and negative forms are checked
individually, since there is always the potential for a name clash with
negative forms on their own, naturally.

of course this test is of no help to out-of-tree plugin development