plugins: add type strictness to macros
This means that type checking is actually performed upon the values
provided by the user, instead of just silently casting things in the
background.

This will be invaluable with the pending capability refactor in the
next commits, which changes one of the types - I lost quite a bit of
time debugging an issue relating to that which came down to having
missed one conversion; this will now cause a compile time error in
such situations, and will be of tremendous value to any 3rd party
out-of-tree plugin developers in the transition to the 4.x API with
the next change.