core: fix audio-misc subcat option prefs placement
as just fixed for video-splitter plugin options in the previous commit,
when a cat/sub-cat is not used in the core option set this can cause
problems for the Qt GUI (at least) organising associated options from
plugins in the tree.

this commit adds a set_subcategory() call for the audio-misc subcat to
the core option set to ensure any options associated with it from
plugins are correctly organised. (though no in-tree plugins with
actual options currently use it).

this subcat has been broken in this way since 2005
(2e3847698a715d877bf4506f4d100ba927ca0696)