core: extend module_config_get() with filter parameters
*note, the old form is still available as a macro!

while most users of module_config_get() do not care about private or
obsolete options, there are use cases that do. this change enables
those use cases to retrieve the data they desire.

one example is extras/analyser/zsh.sh, which currently has a hack of hard
coding in a set of private options; with this change it will no longer
have to and thus will be accurate and uptodate with the proper option set.

another is the unique option check test i just added; this change now
enables that to actually get full access to all the info it needs.