core: don't map plugin where no option callback
the get int/str choices functions already do not map the plugin for the
option of interest where a fixed choice list is given, however they always
otherwise do so, no matter whether or not the option actually has a
callback set for getting a choice list.

this would be fine if only used for options known to have a list, but this
is not the case, for instance see help output where they are used for
every option (of int/str type), and the returned count responded to, which
means help output is significantly less efficient than it should be.

here we simply check that a callback is actually set before doing the work