gui: lookup cat from subcat
having category config records as well as subcat ones is silly, it is
inefficient and bug prone; here we take a step forward to address this,
changing the GUIs to lookup cat from subcat.

this takes the minimal change necessary to rid plugins of set_category()
calls; Qt and MacOSX tree construction are done a little differently from
each other, so while the change here is sufficient also to rid core of them
also on MacOSX, more work in a subseqent commit will be needed for Qt.