modules: do not assume NULL malloc ptr when size is zero
malloc does not necessarily return NULL when given size is zero, thus
we must test for this and return NULL explicitly in the function in
question here.

the documentation of the function did not explicitly claim that a NULL
pointer should be expected when a module has no config items, and some
use cases smartly checked the count instead, but in at least one case
- the QT prefs search code - a NULL pointer was falsely expected.

i encountered a crash there testing another commit, one which happens
to expand the number of plugins with empty option sets (removes those
containing only set cat/subcat entries, no actual options), thus
creating the conditions for the crash to occur. (fyi this flaw is one
of two that allowed this crash to occur, modules with empty option
sets should not even be relevant in the search code, but another flaw
causes the wrong module to be looked at in some cases, which will be
addressed separately).

this commit ensures that this function always returns NULL when the
module has an empty config, and updates the documentation to reflect
this, removing this source of bugs.