gui: use filter mod search in get-filter-type functions
this uses the new capability-specific filtered search ability of name-based
module searches, improving efficiency

one difference in functionality is that we now do a group-by-group search,
rather than returning the very first match across the entire module set,
but typically this is not expected to be a problem - the behaviour of doing
so where there are multiple matches is a problem in itself.

note, ultimately if not found under any of the common types this falls back
to a general search of the entire set; this is only done to maintain the
same behaviour as before with respect to the error issued, but is probably
unnecessary.