modules: skip null-capability mods in mod-list
as noted elsewhere, a handful of "null-capability" modules (those with a
cap of VLC_CAP_INVALID) really do legitimately exist within the module
bank; they exist as a "hack" whereby the first module in a multi-module
plugin is used only for holding a generic name, coming into play when
options are displayed in help output and the GUI preferences interface.

all use of this function (module_list_get()) has been carefully checked,
and such modules are of no interest in all cases, so here we filter them
out, saving callers from having to know to do so.

in terms of callers explicitly filtering them out themselves, most were
wiped out in the introduction and use of module_list_have_config(); the
one that remained has been removed here.

specifically, where this function is used:
 - some users of this function have been recently switched to
   module_list_have_config() and so can be ignored.
 - module_find(), now called module_find_ext(), will be discussed shortly.
 - ListModules() in src/config/help.c is not interested. it prints a list
   of modules, but while this has previously included null-capability
   modules, there is no purpose in doing so; they should be excluded.
 - PluginTab::FillTree() in modules/gui/qt/dialogs/plugins.cpp is not
   interested. it prints a list of modules, and again there is no purpose
   to including them there.
 - both Qt and MacOSX GUIs use it in building lists of modules for module
   selection controls, and have no interest.

with regards to module_find_ext(), this falls back on module_list_get()
when the search is not capability specific:
 - almost all users of the find function are capability specific, and have
   been switched to using its just-added cap-based filter, thus filtering
   out null-capability modules in the list function has no impact on them.
 - the one exception is its use in module_exists(), but all use of that
   function seems to be to find non-null-capability modules, and there is
   no conceiveable use in being able to check null ones.

note, the memory allocated accounts for ALL modules, so there will be some
waste with respect to null-capability ones, but the number is small so it
is not necessarily worth reworking things to gain an accurate allocation.