qt: hide all empty groupboxes
if a panel only has one groupbox and it is empty, it was hidden, but
otherwise they were shown; this was surely a bug, and at any rate is
undesirably messy. all empty groupboxes are now hidden.

an example case: this fixes the empty "Performance options" groupbox (on
my system) under the top-level "Advanced" tree node in the advanced
preferences view, which happens to be very visible as the first panel
shown when switching to advanced mode.