modules: use enums for cat/subcat
this does not at all eliminate all use of cats/subcats as integers (they
are stored in an integer field of config items as a hack), but this at
least has the benefit of enforcing 0..n numbering which we can use (in a
subsequent commit) to make the subcat->cat lookup (and text lookup) much
more efficient.

note, CAT_HIDDEN is deliberately removed at this point as it becomes
unnecessary; all use gets replaced by CAT_INVALID. SUBCAT_HIDDEN is
retained however, as that is still useful.