core: fix advanced-network subcat option prefs placement
as just fixed for video-splitter plugin options in the previous commit,
when a cat/sub-cat is not used in the core option set this can cause
problems for the Qt GUI (at least) organising associated options from
plugins in the tree.

this commit adds a set_subcategory() call for the advanced-network
subcat to the core option set to ensure that options associated with it
from plugins are correctly organised. specifically this affects the
gnu-tls plugin (on my system) which has been misplaced in the tree
ever since e3c92d77d4eb45ac5029f656217b713a1f11685f (2012).