qt: create tree nodes where missing for plugin nodes
the Qt advanced prefs tree construction code was designed to build up a
base cat/subcat tree from the core option set only, then attach nodes for
plugin sets; it was designed on a presumption (i guess) that the core
would always make use of all subcategories itself, thus they would never
be missing when attaching plugin nodes. this presumption is a naive one;
some are not in use and if reference to them is removed completely from
the core, this can have unintended consequences, for instance
5c18cff8d8cdba5784ca4bdb641fb483499117ed removed the reference to the
video-splitter subcat resulting in such plugins being incorrectly attached
to the parent cat node.

the change here results in the function performing a little more work to
build the tree, but it robustifies it against missing subcat references
in the core set. it is also undesirable to have such unused subcat entries
in the core set anyway, since if nothing attaches to them we just end up
with empty nodes.