config: get cat name from general subcat
general subcats represent the options attached directly to a cat node in
the tree. previously, in Qt, the cat name is used for the node label and
the general subcat name is used for the panel title, but, both are
actually the same. macosx just uses the cat name.

it is pointless holding two separate and identical names, one in the cat
table and one in the subcat table.

this gets rid of the name from the cat table, and uses the general subcat's
name through config_CategoryNameGet().

if it were ever the case that different tree-node and panel-title labels
were wanted for any of the base (non-plugin-node) tree, we could/should
introduce a new 'title' attribute to the subcat table along with a
respective get-title accessor function, just like with name.

actually, the 'advanced' cat was alone in actually using a different
title to its label ("Advanced settings" vs "Advanced") and it uses just
"Advanced" for both now, since it's not worth implementing the title
attribute just for that.