config: get cat help from general subcat
following on from doing similar with cat names, when it comes to help text,
this is not even used on macosx (lines commented out), and with Qt, the
help text of general subcats always replaces that of the parent cat, so
effectively the cat help text is unused.

here we remove it from the cat data table and redirect requests to the
respective general subcat.