config: declare a preferred order for categories
currently, in Qt the tree in the advanced preferences view is sorted
entirely alphabetically, thus resulting in the "advanced" category coming
first, when logically it should come last. in macosx i am not certain
whether any sorting is done, it may just be built in the order used in the
core config set.

here we declare the order that we actually want the categories to appear in
(subcats can just be sorted alphabetically under them, and same for plugin
nodes under those).