declink: fix incorrectly hidden option
background: a subcat entry was previously missing before the declaration of
this option; i then added one but left it disabled to not change the
location in the tree where the plugin was attached. now that we have
revised the Qt code so that plugin option sets are correctly partitioned
across multiple nodes in the tree where they specify multiple subcats,
firstly, this means that options appearing before a subcat are no longer
displayed at all and thus this correction is needed now, secondly now
enabling this subcat reference no longer changes the location of all
options for this set, so it's okay to do so now.