macosx: fix potential null deref in prefs tree construction
subCategoryItem would be null if:
 a) a plugin author neglected to place a set_category() call before one
    or more options.
 b) the first or only subcat targetted by a plugin is a general one

there are actual examples of A with a handful of in-tree plugins (to be
fixed in a subsequent commit). i don't have a mac; did it not crash
before this fix?

there are multiple in-tree modules which target general subcats,
including various logger and keystore plugins. again, did this not
cause it to crash?