macosx: fix multi-free unsoundness in prefs tree
whilst most plugins only use a single set_subcategory() call, resulting in
a single node in the prefs tree, a handful use multiple and the code
looked highly suspect to me with regards to handling them correctly.

when the loop processes each plugin, a VLCTreePluginItem object is created,
and holds the copy of the config set returned by a module_config_get()
call. the destructor frees the config set. as the plugin's config set is
iterated over, cat and subcat nodes are created if necessary and the
references updated; for non-hint items, the children of the current
cat/subcat are searched to see if they referenced the VLCTreePluginItem
object, and if not, it was added, thus attaching the plugin to that point
in the tree. leaf nodes are added to the child set of that same
VLCTreePluginItem object.

for multi-subcat plugins, this same VLCTreePluginItem object would be
added to multiple child lists, and effectively there would be multiple
points in the tree which would show that plugins full option set.

the problem comes at destruction time - the VLCTreePluginItem object is
owned by multiple lists and so (i guess, not being hugely familiar with
objective-C) destroyed multiple times, and furthermore this therefore is
freeing the config set object multiple times.

if it were just the config set being free'd multipe times then we could
solve that by just setting the pointer to null (free() is fine with that),
but there's the issue of VLCTreePluginItem destruction.

here we fix the issue by allowing a plugin to only be located in one single
place in the tree, just as has been the case with the Qt plugin. we also
use a simple boolean to track whether or not the node creation has been
done yet, which makes things more efficient.

if we really want the handful of relevant plugins to continue to be located
at multiple points in the tree as before then the solution is to create
fresh VLCTreePluginItem objects for each. however, in a subsequent commit
i am going to revise things so that they indeed show up at multiple
points, but with only the options relevant to each subcat, as just done
with Qt.