macosx: further refactor prefs tree construction
here we introduce a new type - VLCTreeConfigSetOwner - which will hold
ownership of a config set; VLCTreeMainItem now holds an array of these.

thus ownership is moved out of the tree node item types.

the reason for this is to support the next change, which otherwise would
have required duplicate copies of some sets being held.