macosx: refactor prefs tree node types
some plugins want to attach their option set nodes to general subcats, and
thus attach to cat items on the second tier. the existing macosx code did
not support this, with such nodes instead being attached to the previous
non-general subcat node, if available, which is completely wrong.

the problem blocking a simple fix is that the code uses different types
for category/subcategory/plugin nodes, and was designed to expect the three
tiers of branches to map directly to each, not expecting that a mix of
subcat and plugin nodes might be needed on the 2nd tier.

as a solution, here we unify the three branch types into one, with an enum
based property to identify which branch type an instance is. yes, this
wastes a bit of memory, but it's simple and gives us the needed
flexibility.