qt: expand prefs window size a little
(regarding the advanced prefs view)

adjusting the size policy to dedicate more of the width by default to the
panel (the old default was a 50:50 split), helped towards reducing
horizontal scrolling on panels, but the available width is still a little
limited, with several panels still with horizontal scrolling.

a bit more space could shifted from the tree to the panel, but let's try
just bumping up the actual window width a little.

edit: excellent, a great ratio and the only panel with horizontal scrolling
now is the sout/transcode panel, and that's just due to a rather long
name for a single audio filter.