help: use same cat/subcat headings as GUIs
note, there was not a perfect mapping of cats/subcats to cat-hints,
while each cat had a cat-hint, only a few subcats had one. it did
not seem ideal to only implement headings with cats only, ignoring
subcats so I tried honoring both cats and subcats, and although in
a couple of places there are very few options per subcat, it
largely actually works out pretty well and I'm very happy with it.