qt: add expert prefs view
missing from the existing views was any way to know which options were not
in 'default' state, and any way to selectively reset individual options
to 'default' state. the only way to accomplish this was to manually
review/edit the config file, which is not ideal for most users.

this introduces a new 'expert' view (what else to call it?), inspired by
the about:config view in firefox.

like firefox's about:config, we get a complete list of all options (except
GUI-hidden ones of course), with non-default options highlighted in bold,
and a means to reset individual options. there's also a search filter.

note that the existing simple and advanced views are not linked, i.e. a
change to an option in one view is not reflected in the other, and saving
only saves stuff from the current view. similarly, for consistency, this
new expert view is not linked to the other two. imo having these views
separated like this is a flaw, giving poor UX, but it should be fixed
separately.

**TODO:**
editing of individual hotkey items in this new view is currently disabled,
pending figuring out how best to handle it.