modules: do not set longname in set-name callback
module_get_name() falls back on shortname where longname is requested but
is null, and ultimately to object name if both are null (or only shortname
is requested and that is null).

all access to names is done using this function (with one exception in
help.c just fixed).

the copying of object name to longname is undesirable because where-ever
not replaced with something different, we are essentially just wasting
memory wrt saving/loading the plugin cache.