plugins: use correct name setter
module shortnames and longnames are both accessed via module_get_name();
with a fallback on shortname when longname is requested but is null, and
a fallback on object name if shortname is null.

modules setting a longname without also setting a shortname are typicaly
doing things wrong, setting things up for a successful longname request,
but leaving shortname requests having to fall back on simple object names,
when typically the longname is perfectly suited to use as a shortname.

here we simply switch setting longname for setting shortname