help: remodel --list-verbose output
the existing logic simply output the same unordered list of modules as
--list, only with extra lines under each giving 1) capability name and
score, and 2) any shortcuts (one per line). it was done in this way
presumably due to previously lacking any proper organisation of the
modules themselves with which to base a better output on. now though we
have capability specific lists, efficient access to them and a fixed set
of built-in/standard capabilities to iterate over, so we can improve
things.

this:
 1) groups modules by capability
 2) orders them by score (as a natural consequence of this already being
    done in the bank)
 3) ditches the per-module capability and score line, printing score in
    the first line of each module
 4) ditches the 's' for shortcut on the front of shortcut lines

--list output is left as it was

this significant deviation from --list output now actually helps justify
--list-verbose being its own option, rather than the idea that --list
should respect use of -v/--verbose

this new --list-verbose output is much more useful imo

note that the ordering of modules happens to reflect the order in which
they would be tried via module load (not just in terms of score, which is
obvious, but also with respect to those with the same score).