modules: pre-alloc space for objectname
every module and submodule has an object name (stored as the first
shortcut), so move allocation of space for it into the module creation
routine.

this has the effect of turning the VLC_MODULE_NAME action into a simple
setter, rather than creator.

note, doing this now allows add_submodule_name() to be implemented in a
way that utilises the existing VLC_MODULE_NAME action instead of having to
create a new one.