modules: skip null terminator entries of string choice list arrays
normally options should not null terminate arrays for option choice lists,
and in most cases they don't, however, sometimes the array used comes from
a 3rd-party header and is null terminated which presents a problem. it
results in a blank entry in combobox controls, and presents a similar issue
for --help output.

note that in some cases a null is a valid value, thus it is an issue of a
lack of label, and possibly position in the list, whilst in others a null
value is not valid, and thus it must be ignored entirely. examples of the
former includes x264_preset_names, x264_tune_names and x264_profile_names,
while an example of the latter is x264_nal_hrd_names.

this tackles the latter, causing such entries to be ignored. note that it
of course only affects situations where both the value and label arrays
are null terminated, obviously.

**TODO:**
the former case requires arrays to be duplicated with an actual label used
for the null entry. where this is not done, the null value entry will no
longer be made available for use, giving a regression. fixing this is to
be done separately.