core: fix option default inconsistency
the new "expert" Qt preferences view does not understand the special case
that "any" is the same as a NULL or empty string with respect to module
options, and it is undesirable to hack that in. there is thus a bug in that
since most module options specify NULL for their defaults, that if you open
up the modify dialog for such an option, and save without actually changing
the selection (still on "automatic"), the value changes from NULL to "any"
and it is thus marked as modified.

this fixes that bug by switching to using NULL instead of "any" as the
value for the "automatic" entries in module lists.

this also avoids an unnecessary saving of "any" to the saved settings file.